move the main.py file
