+++ depthModels/depthmodels/video_handler.py
@@ -1,0 +1,3 @@
+class VideoHandler:
+    def __init__(self):
+        print("video")
